backend/config.py
import os
from dotenv import load_dotenv
load_dotenv() 


class Config:
    DEBUG = True
    SECRET_KEY = os.getenv('SECRET_KEY', 'default_secret_key')
    
    DB_USER = os.getenv("DB_USER")
    DB_PASS = os.getenv("DB_PASS")
    DB_HOST = os.getenv("DB_HOST")
    DB_PORT = os.getenv("DB_PORT", 5432)
    DB_NAME = os.getenv("DB_NAME")

    SQLALCHEMY_DATABASE_URI = f"postgresql://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
