backend/requirements.txt
alembic==1.17.2
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Limiter==4.0.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
rich==14.2.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==2.0.1
